refactor(core): peel types.ts into domain barrels under types/
Extract board, merge-queue, execution/UI, merge-policy, and workflow-step
domains while keeping types.ts as the browser-safe @fusion/core Vite alias
with local imports so residual interfaces still resolve. Ratchet types.ts
ceiling to live LOC.
